use integers instead of size_t for object identifiers.
Need a better solution for semantics (perl used undef, python would have used None).

+++ src/GUI/Plater/PlaterObject.hpp
@@ -13,7 +13,7 @@ namespace Slic3r { namespace GUI {
 class PlaterObject {
 public:
     std::string name {""};
-    size_t identifier {0U};
+    int identifier {0};
     std::string input_file {""};
     int input_file_obj_idx {-1};
